Refactor: Muevo las expeciones de clientes
- Todas las excepciones que genero en mis Clientes, se propagan, para que la reciba el menuClientes

+++ src/main/java/ar/edu/utn/frbb/tup/service/administracion/clientes/EliminarCliente.java
@@ -1,5 +1,6 @@
 package ar.edu.utn.frbb.tup.service.administracion.clientes;
 
+import ar.edu.utn.frbb.tup.exception.ClienteExistenteException;
 import ar.edu.utn.frbb.tup.model.Cliente;
 import ar.edu.utn.frbb.tup.persistence.ClienteDao;
 import ar.edu.utn.frbb.tup.persistence.CuentaDao;
@@ -24,29 +25,27 @@ public class EliminarCliente extends BaseAdministracion {
     }
 
     //Eliminar cliente
-    public Cliente eliminarCliente(long dni) {
+    public Cliente eliminarCliente(long dni) throws ClienteNoEncontradoException {
 
-        try {
-
-            //Elimino el cliente con el DNI ingresado, si no existe el cliente lanza una excepcion,
-            //devuelve el cliente eliminado para mostrar en pantalla
-            Cliente cliente = clienteDao.deleteCliente(dni);
-
-            //Elimino las relaciones que tiene con las Cuentas y movimientos
-
-            List<Long> CvuEliminar = cuentaDao.getRelacionesDni(dni); //Obtengo lista de todos los CVUs a eliminar
-
-            for (Long cvu : CvuEliminar){
-                cuentaDao.deleteCuenta(cvu);
-                movimientosDao.deleteMovimiento(cvu);
-            }
-
-            return cliente;
-        } catch (ClienteNoEncontradoException ex) {
-            System.out.println("----------------------------------------");
-            System.out.println(ex.getMessage());
-            System.out.println("----------------------------------------");
+        if (clienteDao.findCliente(dni) != null){
+            throw new ClienteNoEncontradoException("Ya existe un cliente con el DNI ingresado");
         }
-        return null;
+
+        //Elimino el cliente con el DNI ingresado, si no existe el cliente lanza una excepcion,
+        //devuelve el cliente eliminado para mostrar en pantalla
+        Cliente cliente = clienteDao.deleteCliente(dni);
+
+
+        //Elimino las relaciones que tiene con las Cuentas y movimientos
+
+        List<Long> CvuEliminar = cuentaDao.getRelacionesDni(dni); //Obtengo lista de todos los CVUs a eliminar
+
+        for (Long cvu : CvuEliminar){
+            cuentaDao.deleteCuenta(cvu);
+            movimientosDao.deleteMovimiento(cvu);
+        }
+
+        return cliente;
     }
+
 }
